Hospitality reception RL env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-healthcare-scheduling"
version = "0.1.0"
description = "Healthcare Appointment Scheduling RL environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.1",
    # Groq SDK for the LLM baseline agent
    "groq>=0.9.0",
    # Pydantic v2 (openenv-core already pins this)
    "pydantic>=2.0.0",
    # ASGI server
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
]

[project.scripts]
# Run the environment server: uv run server
server = "healthcare_scheduling.server.app:main"

[tool.setuptools]
include-package-data = true
packages = [
    "healthcare_scheduling",
    "healthcare_scheduling.server",
    "healthcare_scheduling.tasks",
    "healthcare_scheduling.agent",
]
package-dir = { "healthcare_scheduling" = ".", "healthcare_scheduling.server" = "server", "healthcare_scheduling.tasks" = "tasks", "healthcare_scheduling.agent" = "agent" }
